Replicate #1637
The gist is that we're "forking" our API with a V2 end point. Older
clients can still access the old API, but the newer API is namespaced
differently. Further: the old API is AMS, the newer one is JSON:API.

In tracing the error it seems that Mirage is correctly identifying the
associated model class here:

    https://github.com/samselikoff/ember-cli-mirage/blob/cf90e92842ea5cff23f2a71093af7e3cee89f75f/addon/serializers/json-api-serializer.js#L323

But seems to be tripping up when trying to include linkage data from an
associated model that uses AMS still (which is behaviour we definitely
don't need -- we do need it for the newer JSON:API serializer).